requirements.txt
flask
torch
torchvision
transformers
clip-anytorch
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
pandas==2.1.4
Pillow==10.4.0